style: fix format and split hyperdim_simd.rs to reduce LOC
