+++ src/lib.rs
@@ -77,6 +77,7 @@ pub mod hyperdim;
 mod hyperdim_batch;
 mod hyperdim_serde; // Extracted from hyperdim.rs for LOC gate
 mod hyperdim_simd; // AVX2/NEON SIMD paths
+mod hyperdim_simd_bundle;
 #[cfg(all(not(target_arch = "wasm32"), feature = "mcp"))]
 pub mod mcp;
 pub mod metadata_filter;
